fix logic ssr and csr in upLoadProduct and inventory, feat middleware to block user'access is unauthorized
