+++ FrontEnd/model-trade-app/app/inventory/page.tsx
@@ -1,10 +1,19 @@
 import React from "react";
-import UploadProduct from "../components/inventory/UploadProduct";
+import { cookies } from "next/headers";
+import InventoryPage from "../components/inventory/InventoryPage";
+import { redirect } from "next/navigation";
 
-const page = () => {
+const page = async () => {
+  const cookieStore = await cookies();
+  const token = cookieStore.get("token")?.value;
+
+  // Chỉ kiểm tra token, redirect nếu không có (bổ sung cho Middleware)
+  if (!token) {
+    redirect("/login");
+  }
   return (
     <div>
-      <UploadProduct />
+      <InventoryPage token={token} />
     </div>
   );
 };
